requirements.txt
customtkinter
pyyaml
